Create README.md
Technical Highlights:
Microcontroller Utilization: Leveraged the STM32F103C6 microcontroller for its robust performance and reliability in controlling the robotic arm's movements and operations. This choice was instrumental in achieving precise control and real-time responsiveness.
UART Communication Protocol: Implemented UART communication for seamless data exchange between the STM32 microcontroller and an ESP32 module. This setup was crucial for enabling wireless control and monitoring of the robotic system through a web application.
Web Application Integration: Developed a user-friendly web application that serves as the control interface for the robotic system. This application allows users to remotely operate the robotic arm, including turning the cutting motor on/off and controlling the robot's climbing actions.
Robotics and Automation: The project encapsulates the integration of mechanical design, electronics, and software to create a fully automated solution for climbing palm trees. This includes precise control of the robotic arm for cutting operations and navigating the tree structure